requirements.txt
langchain-core
langchain-openai
langchain-community
langchain-text-splitters
langgraph
python-dotenv
pypdf